delete user works, still in progress

+++ app/controllers/UserController.java
@@ -129,10 +129,10 @@ public class UserController extends Controller {
 		return ok( userList.render(User.all()) );
 	}
 	
-	public static Result delete(String username){
-//		User currentUser = Sesija.getCurrentUser(ctx());
-//		if (currentUser.username.equals(username)
-//				|| Sesija.adminCheck(ctx()))
+	public static Result deleteUser(String username){
+		User currentUser = Sesija.getCurrentUser(ctx());
+		if (currentUser.username.equals(username)
+				|| Sesija.adminCheck(ctx()))
 			User.deleteUser(username);
 		return ok( userList.render(User.all()) );
